Creates HomeViewModel from MainViewModel, separating them for needed functionalities in each view
